refactor: config.py extraction + mypy ratchet (leaf modules strict)
- rocq_mcp/config.py: every server-owned ROCQ_* env knob now has a
  single definition site. server.py re-binds the names (submodules
  read them as _server.<NAME> and tests monkeypatch them there — both
  keep working because server code reads its own globals), so the
  extraction lands with ZERO test retargeting. Names used only
  cross-module carry per-name noqa so ruff's F401 autofix cannot sever
  the attribute paths (it did once — 32 test failures — hence the
  loud comment).
- mypy wired into CI (lint job) and clean over src/: baseline
  check_untyped_defs everywhere, strict (disallow_untyped_defs) on the
  leaf modules (config, taxonomy, envelope, schemas, diag, proof_walk).
  implicit_optional accepted deliberately for the FastMCP
  `ctx: Context = None` idiom.
- The ratchet's first catch: three real hygiene issues in recent code
  (run_assumptions rebinding a variable across types, step_multi's
  summary loop shadowing an except-bound name, unnarrowed line/character
  in closures) — fixed; timeout plumbing annotations in compile.py /
  compile_enrichment.py honestly widened to float end-to-end.

The remaining server.py decomposition (workspace/pet_runtime/app/tools
+ façade, ~115 monkeypatch retargets) stays deliberately deferred: the
plan gates it on first landing/closing the 7 unmerged fork feature
branches, which is a maintainer decision (rebase blast radius).

+++ src/rocq_mcp/compile_enrichment.py
@@ -255,7 +255,7 @@ async def _enrich_compile_failure(
 async def run_compile_with_state(
     source: str,
     workspace: str,
-    timeout: int,
+    timeout: float,
     include_warnings: bool = True,
     lifespan_state: dict[str, Any] | None = None,
 ) -> dict[str, Any]:
@@ -345,7 +345,7 @@ async def _multi_error_walk(
 async def run_compile_file_with_state(
     file: str,
     workspace: str,
-    timeout: int,
+    timeout: float,
     include_warnings: bool = True,
     lifespan_state: dict[str, Any] | None = None,
     keep_vo: bool = False,
